gcp-firebase/examples/python/pyproject.toml
[project]
name = "gcp-firebase-examples"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["dagger-io>=0.19.0"]
